requirements.txt
streamlit
firebase-admin
opencv-python
facenet-pytorch
torch
Pillow
numpy
scikit-learn
requests
